Updated the Production branch to have most of the new features from the LearnOpenGL_Advanced branch, re-written to better fit the codebase and be less hard-coded. Also, some new features were added (mainly a Material and Environment struct). There's a big note in r_renderer.cpp that I'll copy-paste here:
NOTE: This will change almost immediately. I need to decide if I'm sticking with going through a vector of Meshes, switching to a vector of Actors, or something else entirely (JSON? RenderCmds?). I also need to make sure that R_Render doesn't crash if there are no lights, make a better system for acquiring lighting data and using it during the render (i.e: sorting lights to the top of whatever list I use to render), and finally, make the shader work with more than one light (which might mean going back to LearnOpenGL_Advanced for a bit before getting this code un-fucked). However, this code is only hard-coded a little bit and still runs, so I'm okay with pushing it as an update with some temporary solutions that will get changed very soon.